Parametrized types
Estimators and Microdata now take the correlation structure as a parameter to facilitate internal dispatch

+++ src/data/utils.jl
@@ -14,7 +14,7 @@ function adjmsng!(msng::BitVector, corr::Clustered)
     touse     = intersect[corr.msng]
     new_ic    = copy(corr.ic[touse])
     new_nc    = length(unique(new_ic))
-    new_adj   = new_nc / (new_nc - 1)
+    new_adj   = float(new_nc / (new_nc - 1))
     new_mat   = copy(corr.mat[touse, touse])
 
     return Clustered(intersect, new_mat, new_ic, new_adj)
